#730 moving towards solution?
